Videoprocess (videoprocess v0.1.0)

Documentation for Videoprocess.

Wrapper module for ffmpeg providing various video and audio operations

This module only prepares the command, for running, supply the command to Runner

Functions

concat_multimedia(filelist_path, out_path)

ffmpeg function for concatenating video/audio samples

create filelist first (see FileListCreator module)

-y is for always overwrite

Parameters

filelist_path: path to file_list.txt.

out_path: output concatenated video/audio file (e.g. output.ts/output.mp3).

filepath(path, out_folder, extension \\ ".png")

prepare filename accroding to hash and out folder with extension

Parameters

filelist_path: path to file_list.txt.

out_path: output concatenated video/audio file (e.g. output.ts/output.mp3).

input files are: something/1050440-something output is: out_folder/1050440.png

save_audio(video_path, out_dir, extension)

ffmpeg extract and save audio only

Parameters

video_path: path to input video.

out_dir: output directory without filename.

extension: (.mp3|.aac|.wav|.ogg)

save_images(video_file, out_folder, fps)

ffmpeg extract images from video by specified fps

supply the whole video, you may want to use concat function first

save_images_map(opts)

ffmpeg extract images from video by specified fps

supply list of video files

returns list of commands, suitable to be suplied for Runner parallel running

## Parameters Hash containing:

video (list of video files paths)

outfolder

fps

save_video(video_path, out_dir)

ffmpeg extract and save video only

extension matches the input's extension

Parameters

video_path: path to input video.

out_dir: output directory without filename.

save_video(video_path, out_dir, extension)

ffmpeg extract and save video only extension matches the input's extension

Parameters

video_path: path to input video.

out_dir: output directory without filename.

extension: (.mp4|.mkv|.avi|.ts|.h264)

split_multimedia(file_path, out_folder, time)

split input multimedia into segments of length time seconds

Parameters

file_path: path to input video/audio.

out_dir: output directory without filename.

extension: (.mp3|.aac|.wav|.ogg)

wav_to_pcm(file_path)

for PCM sampling